% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BINtools-package.R
\docType{package}
\name{BINtools-package}
\alias{BINtools-package}
\alias{BINtools}
\title{The 'BINtools' package.}
\description{
A DESCRIPTION OF THE PACKAGE
}
\references{
Stan Development Team (2020). RStan: the R interface to Stan. R package version 2.21.2. https://mc-stan.org
}
